requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
passlib
PyJWT
dotenv